app/requirements.txt
python-json-logger==2.0.2
mailru-im-bot==0.0.21
jsonschema==4.6.2
aiohttp==3.8.1
cchardet==2.1.7
aiodns==3.0.0
jinja2==3.1.2
redis==5.0.1
aiohttp-prometheus-client==0.1.2